Bug 755031 - Fix some other warnings in gfx/2d. r=jmuizelaar
UltraBlame original commit: 15500baebbc56cd7b923c048cebdb17c974bd603

+++ gfx/2d/RecordedEvent.cpp
@@ -211,7 +211,9 @@ RecordedEvent::ReadStrokeOptions(std::istream &aStream, StrokeOptions &aStrokeOp
   ReadElement(aStream, aStrokeOptions.mMiterLimit);
   ReadElement(aStream, joinStyle);
   ReadElement(aStream, capStyle);
-  aStrokeOptions.mDashLength = dashLength;
+  
+  
+  aStrokeOptions.mDashLength = size_t(dashLength);
   aStrokeOptions.mLineJoin = joinStyle;
   aStrokeOptions.mLineCap = capStyle;
